requirements/development.txt
-r base.txt
django-debug-toolbar>=1.8
